Generalize Java version deprecation infrastructrue
We will shortly update the minimum Daemon java version to be the same as the 'future' minimum
daemon java version. This means we will no longer emit any deprecations for incompatible
daemon java versions.

We could either git rid of all the infrastructure, or accept that it is still useful even while
we currently have no deprecated java versions. We choose the second case and generalize the
java version deprecation infrastructure, updating it to dynamically generate deprecation messages
based on parameterized java and gradle versions.

Updating the minimum supported Java version for the Gradle daemon should now be as easy as updating
an integer and adding an upgrade guide section.